ONE-4314: Skip migrations for Postgres

+++ lib/migration/src/m20240116_110014_unique_did_in_organisation.rs
@@ -10,6 +10,11 @@ pub const UNIQUE_DID_DID_IN_ORGANISATION_INDEX: &str = "index-Did-Did-Organisati
 #[async_trait::async_trait]
 impl MigrationTrait for Migration {
     async fn up(&self, manager: &SchemaManager) -> Result<(), DbErr> {
+        if manager.get_database_backend() == sea_orm::DatabaseBackend::Postgres {
+            // Skip because it is not supported. If support for Postgres is added in the future
+            // the schema can be setup in its entirety in a new, later migration.
+            return Ok(());
+        }
         manager
             .drop_index(
                 Index::drop()
