+++ lib/migration/src/m20250117_085705_remove_state_tables.rs
@@ -6,6 +6,11 @@ pub struct Migration;
 #[async_trait::async_trait]
 impl MigrationTrait for Migration {
     async fn up(&self, manager: &SchemaManager) -> Result<(), DbErr> {
+        if manager.get_database_backend() == sea_orm::DatabaseBackend::Postgres {
+            // Skip because it is not supported. If support for Postgres is added in the future
+            // the schema can be setup in its entirety in a new, later migration.
+            return Ok(());
+        }
         manager
             .drop_table(Table::drop().table(CredentialState::Table).to_owned())
             .await?;
